sinuweb_hdl_cms/requirements.txt
anyascii==0.3.2
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
defusedxml==0.7.1
Django==5.0.2
django-cors-headers==4.3.1
django-filter==23.5
django-modelcluster==6.2.1
django-permissionedforms==0.1
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.14.0
draftjs-exporter==5.0.0
et-xmlfile==1.1.0
filetype==1.2.0
gunicorn==20.0.4
html5lib==1.1
idna==3.6
l18n==2021.3
laces==0.1.0
openpyxl==3.1.2
pillow==10.2.0
pillow_heif==0.15.0
psycopg==3.1.19
pytz==2024.1
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
telepath==0.3.1
tzdata==2023.4
urllib3==2.2.0
wagtail==6.0
webencodings==0.5.1
Willow==1.8.0
